Add files via upload
Password generator is a handy tool designed to assist users in creating strong and secure passwords. It's built using a combination of CSS, HTML, and JavaScript, each playing a crucial role,
HTML provides the structure of the web page, defining elements like buttons, input fields, and text areas.
CSS is used for styling, enhancing the visual appeal of the interface and ensuring a user-friendly experience.
JavaScript is the backbone of the functionality, responsible for dynamically generating passwords based on user-defined criteria such as length, complexity, and character types.
The project requires a web browser to run, as it's built using standard web technologies. Users can simply open the webpage in their browser to access the password generator. No additional software or plugins are necessary, making it easily accessible to anyone with internet access.
However, it's important to note that while the generated passwords are designed to be strong and secure, users should still exercise caution and follow best practices for password management, such as avoiding the reuse of passwords across multiple accounts and enabling two-factor authentication where possible.